перевод каталога товаров и корзины на thymeleaf

+++ src/main/java/org/example/webApplicationShopSpringBoot/dto/dto/OrderDTO.java
@@ -17,7 +17,7 @@ public class OrderDTO {
 
     private Long productId;
 
-    private Integer Count;
+    private Long Count;
 
     private BigDecimal productPrice;
