+++ src/main/java/org/example/webApplicationShopSpringBoot/service/userOrderProduct/UserOrderProductService.java
@@ -2,5 +2,5 @@ package org.example.webApplicationShopSpringBoot.service.userOrderProduct;
 
 public interface UserOrderProductService {
 
-    void changeProductCount(Integer UserOrderId, Integer productId, Integer count);
+    void changeProductCount(Long UserOrderId, Long productId, Long count);
 }
